pyproject.toml
[tool.poetry]
name = "fastapi_sqlalchemy"
version = "0.1.0"
description = ""
authors = ["Sviatoslav Kovaliv <sviatoslav.kovaliv@wesoftyou.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.73.0"
uvicorn = {extras = ["standard"], version = "^0.17.1"}
aiosqlite = "^0.17.0"
sqlmodel = "^0.0.6"
alembic = "^1.7.5"
confz = "^1.3.0"
PyJWT = "^2.3.0"
passlib = "^1.7.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
